+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-# [1.0.1-alpha.5] - 10-04-2023
+# [1.0.1-alpha.5] 
 ## [Unreleased]
 This alpha has Bug fixes and minor changes.